+++ src/cript/api/__init__.py
@@ -1,5 +1,6 @@
 # trunk-ignore-all(ruff/F401)
 
 from cript.api.api import API
+from cript.api.data_schema import DataSchema
 from cript.api.valid_search_modes import SearchModes
 from cript.api.vocabulary_categories import VocabCategories
