+++ tests/api/test_api.py
@@ -11,9 +11,7 @@ import requests
 
 import cript
 from conftest import HAS_INTEGRATION_TESTS_ENABLED
-from cript.api.exceptions import InvalidVocabulary
 from cript.api.paginator import Paginator
-from cript.nodes.exceptions import CRIPTNodeSchemaError
 
 
 def test_api_with_invalid_host() -> None:
@@ -84,174 +82,6 @@ def test_config_file() -> None:
 
         assert api._host == config_file_texts["host"] + "/api/v1"
         assert api._api_token == config_file_texts["api_token"]
-
-
-@pytest.mark.skip(reason="too early to write as there are higher priority tasks currently")
-def test_api_initialization_stress() -> None:
-    """
-    tries to put the API configuration under as much stress as it possibly can
-    it tries to give it mixed options and try to trip it up and create issues for it
-
-    ## scenarios
-    1. if there is a config file and other inputs, then config file wins
-    1. if config file, but is missing an attribute, and it is labeled as None, then should get it from env var
-    1. if there is half from input and half from env var, then both should work happily
-    """
-    pass
-
-
-def test_get_db_schema_from_api(cript_api: cript.API) -> None:
-    """
-    tests that the Python SDK can successfully get the db schema from API
-    """
-    db_schema = cript_api._get_db_schema()
-
-    assert bool(db_schema)
-    assert isinstance(db_schema, dict)
-
-    # db schema should have at least 30 fields
-    assert len(db_schema["$defs"]) > 30
-
-
-def test_is_node_schema_valid(cript_api: cript.API) -> None:
-    """
-    test that a CRIPT node can be correctly validated and invalidated with the db schema
-
-    * test a couple of nodes to be sure db schema validation is working fine
-        * material node
-        * file node
-    * test db schema validation with an invalid node, and it should be invalid
-
-    Notes
-    -----
-    * does not test if serialization/deserialization works correctly,
-    just tests if the node schema can work correctly if serialization was correct
-
-    # TODO the tests here only test POST db schema and not PATCH yet, those tests must be added
-    """
-
-    # ------ invalid node schema------
-    invalid_schema = {"invalid key": "invalid value", "node": ["Material"]}
-
-    with pytest.raises(CRIPTNodeSchemaError):
-        cript_api._is_node_schema_valid(node_json=json.dumps(invalid_schema), is_patch=False)
-
-    # ------ valid material schema ------
-    # valid material node
-    valid_material_dict = {"node": ["Material"], "name": "0.053 volume fraction CM gel", "uid": "_:0.053 volume fraction CM gel"}
-
-    # convert dict to JSON string because method expects JSON string
-    assert cript_api._is_node_schema_valid(node_json=json.dumps(valid_material_dict), is_patch=False) is True
-    # ------ valid file schema ------
-    valid_file_dict = {
-        "node": ["File"],
-        "source": "https://criptapp.org",
-        "type": "calibration",
-        "extension": ".csv",
-        "data_dictionary": "my file's data dictionary",
-    }
-
-    # convert dict to JSON string because method expects JSON string
-    assert cript_api._is_node_schema_valid(node_json=json.dumps(valid_file_dict), is_patch=False) is True
-
-
-def test_is_node_schema_valid_skipped(cript_api: cript.API) -> None:
-    """
-    test that a CRIPT node can be correctly validated and invalidated with the db schema, when skipping tests is active
-
-    * test db schema validation with an invalid node, and it should be invalid, but only detected if forced
-
-    Notes
-    -----
-    * does not test if serialization/deserialization works correctly,
-    just tests if the node schema can work correctly if serialization was correct
-
-    """
-
-    def extract_base_url(url):
-        # Split the URL by "//" first to separate the scheme (like http, https)
-        parts = url.split("//", 1)
-        scheme, rest = parts if len(parts) > 1 else ("", parts[0])
-
-        # Split the rest by the first "/" to separate the domain
-        domain = rest.split("/", 1)[0]
-        return f"{scheme}//{domain}" if scheme else domain
-
-    with cript.API(host=extract_base_url(cript_api.host), api_token=cript_api._api_token, storage_token=cript_api._storage_token) as local_cript_api:
-        local_cript_api.skip_validation = True
-        # ------ invalid node schema------
-        invalid_schema = {"invalid key": "invalid value", "node": ["Material"]}
-
-        # Test should be skipped
-        assert local_cript_api._is_node_schema_valid(node_json=json.dumps(invalid_schema), is_patch=False) is None
-
-        with pytest.raises(CRIPTNodeSchemaError):
-            local_cript_api._is_node_schema_valid(node_json=json.dumps(invalid_schema), is_patch=False, force_validation=True)
-
-
-def test_get_vocabulary_by_category(cript_api: cript.API) -> None:
-    """
-    tests if a vocabulary can be retrieved by category
-    1. tests response is a list of dicts as expected
-    1. create a new list of just material identifiers
-    1. tests that the fundamental identifiers exist within the API vocabulary response
-
-    Warnings
-    --------
-    This test only gets the vocabulary category for "material_identifier_key" and does not test all the possible
-    CRIPT controlled vocabulary
-    """
-
-    material_identifier_vocab_list = cript_api.get_vocab_by_category(cript.VocabCategories.MATERIAL_IDENTIFIER_KEY)
-
-    # test response is a list of dicts
-    assert isinstance(material_identifier_vocab_list, list)
-
-    material_identifiers = [identifier["name"] for identifier in material_identifier_vocab_list]
-
-    # assertions
-    assert "bigsmiles" in material_identifiers
-    assert "smiles" in material_identifiers
-    assert "pubchem_cid" in material_identifiers
-
-
-def test_get_controlled_vocabulary_from_api(cript_api: cript.API) -> None:
-    """
-    checks if it can successfully get the controlled vocabulary list from CRIPT API
-    """
-    number_of_vocab_categories = 26
-    vocab = cript_api._get_vocab()
-
-    # assertions
-    # check vocabulary list is not empty
-    assert bool(vocab) is True
-    assert len(vocab) == number_of_vocab_categories
-
-
-def test_is_vocab_valid(cript_api: cript.API) -> None:
-    """
-    tests if the method for vocabulary is validating and invalidating correctly
-
-    * test with custom key to check it automatically gives valid
-    * test with a few vocabulary_category and vocabulary_words
-        * valid category and valid vocabulary word
-    * test that invalid category throws the correct error
-        * invalid category and valid vocabulary word
-    * test that invalid vocabulary word throws the correct error
-        * valid category and invalid vocabulary word
-    tests invalid category and invalid vocabulary word
-    """
-    # custom vocab
-    assert cript_api._is_vocab_valid(vocab_category=cript.VocabCategories.ALGORITHM_KEY, vocab_word="+my_custom_key") is True
-
-    # valid vocab category and valid word
-    assert cript_api._is_vocab_valid(vocab_category=cript.VocabCategories.FILE_TYPE, vocab_word="calibration") is True
-    assert cript_api._is_vocab_valid(vocab_category=cript.VocabCategories.QUANTITY_KEY, vocab_word="mass") is True
-    assert cript_api._is_vocab_valid(vocab_category=cript.VocabCategories.UNCERTAINTY_TYPE, vocab_word="fwhm") is True
-
-    # valid vocab category but invalid vocab word
-    with pytest.raises(InvalidVocabulary):
-        cript_api._is_vocab_valid(vocab_category=cript.VocabCategories.FILE_TYPE, vocab_word="some_invalid_word")
 
 
 def test_download_file_from_url(cript_api: cript.API, tmp_path) -> None:
